[#33] Refactor split method to its own class
